+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed "Catch Critter" challenge to spawn 2 critters nearby, closer than before.
 - Changed "Catch Critter" challenge to highlight on-screen critters.
 - Changed "Craft Item" challenge to accept any torch or wooden sword.
+- Changed "Stand On Block" challenge to accept any wood.
 
 ## [2.0.6] - 2023-09-29